==> 1_scraper.ipynb <==
import pandas as pd
import numpy as np

#the website blocks the url-agent used by urllib
#and I can't find a way to change it
#so currently this code block is useless

'''from urllib.request import urlretrieve
from bs4 import BeautifulSoup
import requests

bulk_dl_url = 'https://www.federalregister.gov/presidential-documents/executive-orders'
link_prefix = 'https://www.federalregister.gov/'
r = requests.get(bulk_dl_url)
html_doc = r.text
soup = BeautifulSoup(html_doc)
dls = soup.find(class_='bulk-files')
dls_all = dls.find_all('li')
links = []
names = []
for dl in dls_all:
    names.append(dl.span.text)
    link = dl.find('a')
    links.append(link_prefix + link['href'])
links.pop(0)
names.pop(0)

names = [name.replace('.', '') for name in names]
names = [name.replace(' ', '_') for name in names]
files = [name + '.csv' for name in names]

for i, link in enumerate(links):
    urlretrieve(link, names[i])'''

# step 1: get the data

# after 1994, orders are available as a bulk csv or json download at
# https://www.federalregister.gov/documents/search.csv?conditions%5Bcorrection%5D=0&conditions%5Bpresident%5D=&conditions%5Bpresidential_document_type%5D=executive_order&conditions%5Bsigning_date%5D%5Byear%5D=&conditions%5Btype%5D%5B%5D=PRESDOCU&fields%5B%5D=citation&fields%5B%5D=document_number&fields%5B%5D=end_page&fields%5B%5D=html_url&fields%5B%5D=pdf_url&fields%5B%5D=type&fields%5B%5D=subtype&fields%5B%5D=publication_date&fields%5B%5D=signing_date&fields%5B%5D=start_page&fields%5B%5D=title&fields%5B%5D=disposition_notes&fields%5B%5D=executive_order_number&order=executive_order&per_page=1000
# so this will only scrape 1937 (when the register starts) through 1994

# orders are on a list of pages, denoted by year (or year-president for inauguration years)
#base_url = 'https://www.archives.gov/federal-register/executive-orders/'
# likely not the most elegant way of doing it, but it gets the job done
years = [*range(1937, 1995)]
inaug_years = [1945, 1953, 1961, 1963, 1969, 1974, 1977, 1981, 1989, 1993]
presidents = ['roosevelt', 'truman', 'eisenhower', 'kennedy', 'johnson',
             'nixon', 'ford', 'carter', 'reagan', 'bush', 'clinton']
years = [str(y) for y in years]
inaug_years = [str(iy) for iy in inaug_years]

year_pres = []
i=0
# create the suffixes for inauguration years
for year in inaug_years:
    year_pres.append(year+'-'+presidents[i])
    i += 1
    year_pres.append(year+'-'+presidents[i])

years = list(set(years) - set(inaug_years))
years = years + year_pres
print(years)
years.sort()
print(years)

def scraper(url):
    r = requests.get(url)
    html_doc = r.text
    soup = BeautifulSoup(html_doc)
    mainblock = soup.div.find(id = 'block-system-main')
    subs = mainblock.find_all(['p', 'ul'])

    this_list = [sub.text for sub in subs]

    df = pd.DataFrame(this_list, columns = ['data'])
    return(df)

base_url = 'https://www.archives.gov/federal-register/executive-orders/'
df_all = pd.DataFrame(columns = ['data'])
for year in years:
    url = base_url + year
    df_all = df_all.append(scraper(url), ignore_index = True)
    print(year)

print(df_all.shape)

#strip leading and trailing whitespaces for easier processing
df_all['data'] = df_all['data'].str.strip()

#replace empty strings and recurring non-data with NaN to drop
df_all = df_all.replace(r'^\s*$', np.NaN, regex=True)
df_all = df_all.replace('Top of Page', np.NaN, regex=False)
df_all = df_all.replace('Ronald Reagan Disposition Tables', np.NaN, regex=False)
df_all.dropna(axis = 0, how = 'all', inplace = True)

print(df_all.shape)
df_all.to_csv('scraped_data.csv', index = False)
