==> 2_clean_data.ipynb <==
#step 2: clean the data because even though it's regarding the highest office in the country,
#it's still a govennment website

import pandas as pd
import numpy as np
pd.options.mode.chained_assignment = None

df_old = pd.read_csv('scraped_data.csv')
print(df_old.shape)
print(df_old.head())

#This is the ugly part where I have to manually edit some parts of the data
#pls forgive me

#just to see the rows of data that need manual edits
df_bads = df_old[~df_old['data'].str.startswith('Executive')]
df_bads = df_bads[~df_bads['data'].str.startswith('Signed')]
print(df_bads.shape)
print(df_bads)

#lil ugly manual adjustments to insert missing order numbers
#row 9582 - order number 12287 - 1981
df_old.loc[[9582]] = df_old.loc[[9582]].replace('Decontrol of crude oil and refined petroleum products', 'Executive Order 12287 \nDecontrol of crude oil and refined petroleum products', regex=False)

#row 9808 - order number 12400 - 1983
df_old.loc[[9808]] = df_old.loc[[9808]].replace('President\'s Commission on Strategic Forces', 'Executive Order 12400 \nPresident\'s Commission on Strategic Forces', regex=False)

#row 10004 - order number 12498 - 1985
df_old.loc[[10004]] = df_old.loc[[10004]].replace('Regulatory planning process', 'Executive Order 12498 \nRegulatory planning process', regex=False)

#row 10094 - order number 12543 - 1986
df_old.loc[[10094]] = df_old.loc[[10094]].replace('Prohibiting trade and certain transactions involving Libya', 'Executive Order 12543 \nProhibiting trade and certain transactions involving Libya', regex=False)

#row 10254 - order number 12623 - 1988
df_old.loc[[10254]] = df_old.loc[[10254]].replace('Delegating authority to implement assistance to the Nicaraguan Democratic Resistance', 'Executive Order 12623 \nDelegating authority to implement assistance to the Nicaraguan Democratic Resistance', regex=False)

#row 10396 - order number 12663 - 1989
df_old.loc[[10396]] = df_old.loc[[10396]].replace('Visit the Ronald Reagan Presidential Library and Museum', 'Executive Order 12663 \nAdjustments of certain rates of pay and allowances', regex=False)

#fixing formatting/typos of a few signing dates
#row 2145 - signing date and fed register page and date are mashed together
df_old.loc[[2145]] = df_old.loc[[2145]].replace('Signed: 5 FR 4558, November 18, 194November 20, 1940', 'Signed: November 18, 1940 \nFederal Register page and date: 5 FR 4558, November 20, 1940', regex=True)

#row 1533 - insert '\n' after signing date
df_old.loc[[1533]] = df_old.loc[[1533]].replace('Signed: November 30, 1939 ', 'Signed: November 30, 1939 \n', regex=True)

#row 5201 - change year from 950 to 1950
df_old.loc[[5201]] = df_old.loc[[5201]].replace('Signed: January 31, 950', 'Signed: January 31, 1950', regex=True)

#row 6243 - insert space between comma and year
df_old.loc[[6243]] = df_old.loc[[6243]].replace('Signed: June 29,1955', 'Signed: June 29, 1955', regex=True)

#row 6367 - correct year from 19656 to 1956
df_old.loc[[6367]] = df_old.loc[[6367]].replace('Signed: October 22, 19656', 'Signed: October 22, 1956', regex=True)

#row 4871 - correct year from 1945 to 1948
df_old.loc[[4871]] = df_old.loc[[4871]].replace('Signed: March 22, 1945', 'Signed: March 22, 1948', regex=True)

#row 4563 - correct year from 1948 to 1946
df_old.loc[[4563]] = df_old.loc[[4563]].replace('Signed: October 5, 1948', 'Signed: October 5, 1946', regex=True)

#row 2893 - correct year from 1946 to 1941
df_old.loc[[2893]] = df_old.loc[[2893]].replace('Signed: December 11, 1946', 'Signed: December 11, 1941', regex=True)

#row 5551 - correct year from 1961 to 1951
df_old.loc[[5551]] = df_old.loc[[5551]].replace('Signed: July 31, 1961', 'Signed: July 31, 1951', regex=True)

#row 5551 - correct date from July 6 to July 26
df_old.loc[[2619]] = df_old.loc[[2619]].replace('Signed: July 6, 1941', 'Signed: July 26, 1941', regex=True)

#row 5551 - correct date from August 1 to August 11
df_old.loc[[5301]] = df_old.loc[[5301]].replace('Signed: August 1, 1950', 'Signed: August 11, 1950', regex=True)

print(df_old.loc[[5301]])

#to view and verify they updated properly
print(df_old.loc[[9582]])
print(df_old.loc[[9808]])
print(df_old.loc[[10004]])
print(df_old.loc[[10094]])
print(df_old.loc[[10254]])
print(df_old.loc[[10396]])
print(df_old.loc[[2145]])
print(df_old.loc[[1533]])
print(df_old.loc[[5201]])
print(df_old.loc[[6243]])
print(df_old.loc[[6367]])
print(df_old.loc[[4871]])
print(df_old.loc[[4563]])
print(df_old.loc[[2893]])
print(df_old.loc[[5551]])
print(df_old.loc[[2619]])
print(df_old.loc[[5301]])

#split the dataframe and make sure they're of equal length
df_nums = df_old[df_old['data'].str.startswith('Executive')]
df_dates = df_old[df_old['data'].str.startswith('Signed')]
print(df_nums.shape)
print(df_dates.shape)

#trim unecessary strings and make for easier processing
#because some rows have '\n' between 'Executive Order' and the order number
df_nums['data'] = df_nums['data'].str.replace('Executive Order', '', regex=False)
df_nums['data'] = df_nums['data'].str.replace('No.', '', regex=False)
df_nums['data'] = df_nums['data'].str.strip()
#split the order number and title
df_nums[['executive_order_number', 'title']] = df_nums['data'].str.split('\n', n=1, expand=True)

print(df_nums.head())

df_dates['data'] = df_dates['data'].str.replace('Signed:', '', regex=False)
df_dates['data'] = df_dates['data'].str.strip()
#split signing date and disposition notes
df_dates[['signing_date', 'disposition_notes']] = df_dates['data'].str.split('\n', n=1, expand=True)

print(df_dates.head())

df_nums['count'] = [*range(df_nums.shape[0])]
df_dates['count'] = [*range(df_dates.shape[0])]
df_all = df_nums[['executive_order_number', 'title', 'count']].merge(df_dates[['signing_date', 'disposition_notes', 'count']], on='count')
df_all = df_all.drop(['count'], axis = 1)

for col in list(df_all):
    df_all[col] = df_all[col].str.strip()

print(df_all.head())

df_all['executive_order_number'] = df_all['executive_order_number'].str.replace('-A', '.2', regex = False)
df_all['executive_order_number'] = df_all['executive_order_number'].str.replace('A', '.5', regex = False)
df_all['executive_order_number'] = df_all['executive_order_number'].str.replace('-B', '.5', regex = False)
df_all['executive_order_number'] = df_all['executive_order_number'].str.replace('-C', '.7', regex = False)
df_all['executive_order_number'] = df_all['executive_order_number'].astype(float)
df_all.sort_values(by='executive_order_number', ascending=False, inplace = True)

df_all.to_csv('cleaned_data.csv', index=False)
