==> 3_combine_data.ipynb <==
import pandas as pd
import datetime
pd.options.mode.chained_assignment = None

#import data from bulk CSV downloads

file_prefix = 'documents_signed_by_'
pres_names = ['joseph_r_biden_jr', 'donald_trump', 'barack_obama', 'george_w_bush', 'william_j_clinton']
file_suffix = '_of_type_presidential_document_and_of_presidential_document_type_executive_order.csv'
colnames = ['executive_order_number', 'title', 'signing_date', 'disposition_notes']

df_all = pd.DataFrame(columns = colnames)
for name in pres_names:
    filename = file_prefix + name + file_suffix
    df = pd.read_csv(filename)
    df_subset = df[colnames]
    print(name)
    df_all = df_all.append(df_subset, ignore_index = True)
print(df_all.head())
print(df_all.shape)

df_old = pd.read_csv('cleaned_data.csv')
print(df_old.shape)
print(df_all.shape)
df_all = df_all.append(df_old, ignore_index = True)
df_all['executive_order_number'] = df_all['executive_order_number'].astype(float)
df_all.sort_values(by='executive_order_number', inplace = True)

print(df_all.shape)

#manual edit for error in Obama bulk csv
print(df_all['signing_date'].loc[[514]])
df_all['signing_date'].loc[[514]] = df_all['signing_date'].loc[[514]].replace('12/29/2010', '12/29/2009', regex=False)
print(df_all['signing_date'].loc[[514]])

#convert dates to datetime for calculations

df_all['signing_date'] = pd.to_datetime(df_all['signing_date'])

#add columns for name, party, days in office, and count
df_all['president'] = 'name'
df_all['party'] = 'party'
df_all['days'] = 0
df_all['count'] = 0
print(df_all.shape)

#import inaug dates with csv
df_inaug = pd.read_csv('inaug_dates.csv')
today = datetime.date.today()
print(today)
df_inaug.loc[len(df_inaug.index)] = ['Today', today, 'party']

df_inaug['inaug_date'] = pd.to_datetime(df_inaug['inaug_date'])
print(df_inaug)

#subset df by dates
for i in range(df_inaug.shape[0] - 1):
    name = df_inaug.at[i, 'name']
    start_date = df_inaug.at[i, 'inaug_date']
    end_date = df_inaug.at[i+1, 'inaug_date']

    df_subset = df_all.loc[(df_all['signing_date'] >= start_date) & (df_all['signing_date'] < end_date)]
    df_subset['president'] = df_inaug.at[i, 'name']
    df_subset['party'] = df_inaug.at[i, 'party']
    df_subset['days'] = df_subset['signing_date'] - start_date
    df_subset['count'] = [*range(1, df_subset.shape[0] + 1)]
    
    #have to update for FDR, data starts partway through his presidency
    if i == 0:
        df_subset['count'] = df_subset['count'] + 1715
    df_all.update(df_subset)
    print(name, start_date, end_date)
    print(df_subset.shape)
    
print(df_all.head())

df_all['days'] = df_all['days'].dt.days
print(df_all.head())

df_all.to_csv('all_data.csv', index=False)
