==> 3_pretty_graphs.ipynb <==
#step 3: uhhhhhhhhhhh
#step 4: pretty graph?
#step 5: interesting stats?

#see which orders revoke others (who revokes who? who revokes most per orders signed/per year?)
#which orders supercede others? (who supercedes who? who supercedes most per orders signed/per year?)
