==> 5_further_analysis.ipynb <==
#step 5: more analysis

#see which orders revoke others (who revokes who? who revokes most per orders signed/per year? who was most revoked per signed/total?)
#Revoke, Amend, Supercede

import pandas as pd
df = pd.read_csv('all_data.csv')
print(df.shape)
df = df.dropna(subset=['disposition_notes'])
print(df.shape)

df_revoke = df[df['disposition_notes'].str.contains('Revoke')]
print(df_revoke.shape)

df_revokes = df[df['disposition_notes'].str.contains('Revokes')]
print(df_revokes.shape)

df_revoked = df[df['disposition_notes'].str.contains('Revoked')]
print(df_revoked.shape)

df_both = df[(df['disposition_notes'].str.contains('Revokes')) & (df['disposition_notes'].str.contains('Revoked'))]
print(df_both.shape)
